feat(stock): add stock feature
Added stock feature to product model. Handeled order creation on condition of stock availibility

+++ sasta_amazon/order/views.py
@@ -8,20 +8,27 @@ from django.contrib.auth.decorators import login_required
 def creat_order(request):
     user=request.user
     cart=Cart.objects.get(user=user)
-
     order=Order.objects.create(user=user)
+    created=False
 
     for item in cart.items.all():
-        OrderItem.objects.create(
-            order=order,
-            product=item.product,
-            quantity=item.quantity,
-            price=item.product.price
-            )
-
+        if item.product.stock>=item.quantity:
+            OrderItem.objects.create(
+                order=order,
+                product=item.product,
+                quantity=item.quantity,
+                price=item.product.price
+                )
+            item.product.stock-=item.quantity
+            item.product.save()
+            created=True
+    if created:
         cart.items.all().delete()
+        return redirect('success_order',order_id=order.id)
+    else:
+        return redirect('cart')
+
     
-    return redirect('success_order',order_id=order.id)
 
 @login_required    
 def success_order(request,order_id):
